requirements_simple.txt
requests>=2.25.0
python-dotenv>=0.19.0
keyring>=23.0.0
